# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Office management control system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-logs:
    get:
      operationId: listWorkLogs
      tags: [work-logs]
      parameters:
        - in: query
          name: staffId
          schema: { type: integer }
        - in: query
          name: taskId
          schema: { type: integer }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: Work logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkLog" }
    post:
      operationId: createWorkLog
      tags: [work-logs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWorkLogBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkLog" }
  /work-logs/{id}:
    patch:
      operationId: updateWorkLog
      tags: [work-logs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWorkLogBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkLog" }
    delete:
      operationId: deleteWorkLog
      tags: [work-logs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Invoices ----------
...
components:
  schemas:
...
    WorkLog:
      type: object
      required: [id, staffId, staffName, date, summary, createdAt]
      properties:
        id: { type: integer }
        staffId: { type: integer }
        staffName: { type: string }
        taskId: { type: integer, nullable: true }
        taskTitle: { type: string, nullable: true }
        date: { type: string, format: date }
        summary: { type: string }
        hours: { type: number, nullable: true }
        status:
          type: string
          enum: [in_progress, blocked, completed]
        createdAt: { type: string, format: date-time }
    CreateWorkLogBody:
      type: object
      required: [date, summary, status]
      properties:
        taskId: { type: integer, nullable: true }
        date: { type: string, format: date }
        summary: { type: string, minLength: 1 }
        hours: { type: number, nullable: true }
        status:
          type: string
          enum: [in_progress, blocked, completed]
    UpdateWorkLogBody:
      type: object
      properties:
        taskId: { type: integer, nullable: true }
        date: { type: string, format: date }
        summary: { type: string }
        hours: { type: number, nullable: true }
        status:
          type: string
          enum: [in_progress, blocked, completed]
...